weston_rootfs: conditionally install PM_UTILS scripts

+++ variscite/weston_rootfs.sh
@@ -203,21 +203,24 @@ function rootfs_install_config_weston_service() {
 }
 
 function rootfs_install_config_pm_utils() {
-	# remove pm-utils default scripts and install wifi / bt pm-utils script
-	rm -rf ${ROOTFS_BASE}/usr/lib/pm-utils/sleep.d/
-	rm -rf ${ROOTFS_BASE}/usr/lib/pm-utils/module.d/
-	rm -rf ${ROOTFS_BASE}/usr/lib/pm-utils/power.d/
-	install -m 0755 ${PM_UTILS_DIR}/01-bt.sh \
-		${ROOTFS_BASE}/etc/pm/sleep.d/
-	install -m 0755 ${PM_UTILS_DIR}/02-wifi.sh \
-		${ROOTFS_BASE}/etc/pm/sleep.d/
-	if [ -f ${PM_UTILS_DIR}/03-eth.sh ]; then
-		install -m 0755 ${PM_UTILS_DIR}/03-eth.sh \
+	# install freertos-variscite
+	if [ ! -z "${PM_UTILS_DIR}" ]; then
+		# remove pm-utils default scripts and install wifi / bt pm-utils script
+		rm -rf ${ROOTFS_BASE}/usr/lib/pm-utils/sleep.d/
+		rm -rf ${ROOTFS_BASE}/usr/lib/pm-utils/module.d/
+		rm -rf ${ROOTFS_BASE}/usr/lib/pm-utils/power.d/
+		install -m 0755 ${PM_UTILS_DIR}/01-bt.sh \
 			${ROOTFS_BASE}/etc/pm/sleep.d/
-	fi
-	if [ -f ${PM_UTILS_DIR}/04-usb.sh ]; then
-		install -m 0755 ${PM_UTILS_DIR}/04-usb.sh \
+		install -m 0755 ${PM_UTILS_DIR}/02-wifi.sh \
 			${ROOTFS_BASE}/etc/pm/sleep.d/
+		if [ -f ${PM_UTILS_DIR}/03-eth.sh ]; then
+			install -m 0755 ${PM_UTILS_DIR}/03-eth.sh \
+				${ROOTFS_BASE}/etc/pm/sleep.d/
+		fi
+		if [ -f ${PM_UTILS_DIR}/04-usb.sh ]; then
+			install -m 0755 ${PM_UTILS_DIR}/04-usb.sh \
+				${ROOTFS_BASE}/etc/pm/sleep.d/
+		fi
 	fi
 }
